backend/requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-retry==2.9.1
aiosignal==1.4.0
aiosqlite==0.22.1
alembic==1.18.4
altgraph==0.17.4
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.94.0
anyio==4.12.0
attrs==25.3.0
bcrypt==4.0.1
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.2
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.7
cycler==0.12.1
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
ecdsa==0.19.2
email-validator==2.3.0
fastapi==0.136.0
flatbuffers==25.2.10
fonttools==4.60.0
frozenlist==1.8.0
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
jax==0.7.2
jaxlib==0.7.2
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.9
langchain-anthropic==1.4.0
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.28
langchain-openai==1.1.6
langchain-pinecone==0.2.13
langchain-text-splitters==1.1.0
langgraph==1.1.6
langgraph-checkpoint==4.0.1
langgraph-checkpoint-sqlite==3.0.3
langgraph-prebuilt==1.0.9
langgraph-sdk==0.3.13
langsmith==0.5.1
Mako==1.3.11
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib==3.10.6
mediapipe==0.10.14
ml_dtypes==0.5.3
multidict==6.7.0
mypy_extensions==1.1.0
numpy==2.2.6
openai==2.14.0
opencv-contrib-python==4.12.0.88
opencv-python==4.12.0.88
opt_einsum==3.4.0
orjson==3.11.5
ormsgpack==1.12.2
packaging==24.2
passlib==1.7.4
pefile==2023.2.7
pillow==11.3.0
pinecone==7.3.0
pinecone-plugin-assistant==1.8.0
pinecone-plugin-interface==0.0.7
propcache==0.4.1
protobuf==4.25.8
psycopg2-binary==2.9.11
pyasn1==0.6.3
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pygame==2.6.1
pyinstaller==6.15.0
pyinstaller-hooks-contrib==2025.8
pyparsing==3.2.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.26
pywin32-ctypes==0.2.3
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
scipy==1.16.2
setuptools==80.9.0
simsimd==6.5.12
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.2
SQLAlchemy==2.0.45
sqlite-vec==0.1.9
starlette==1.0.0
tenacity==9.1.2
tiktoken==0.12.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
uuid_utils==0.12.0
uv==0.9.11
uvicorn==0.44.0
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
